Remove all decoders minus libvips
