--- library/src/main/cpp/image-decoder/decoder_webp.h
@@ -1,25 +1,0 @@
-//
-// Created by len on 30/12/20.
-//
-
-#ifndef IMAGEDECODER_DECODER_WEBP_H
-#define IMAGEDECODER_DECODER_WEBP_H
-
-#include "decoder_base.h"
-#include <webp/decode.h>
-#include <webp/demux.h>
-
-class WebpDecoder : public BaseDecoder {
-public:
-  WebpDecoder(std::shared_ptr<Stream>&& stream, bool cropBorders,
-              cmsHPROFILE targetProfile);
-
-  void decode(uint8_t* outPixels, Rect outRect, Rect inRect,
-              uint32_t sampleSize);
-
-private:
-  ImageInfo parseInfo();
-  cmsHPROFILE getColorProfile();
-};
-
-#endif // IMAGEDECODER_DECODER_WEBP_H
